Rewrote motivation to index
Per suggestions in PR #9, the motivation.md file content has been moved to the index.md file and rewritten to both be more document storage-format agnostic and place an empathis on digidal garden creation earlier on. 

I also expanded it with some general ideas on the vision of where NoteMate could be expanded to.