Documentation coverage (#9)
* Initial documentation bits

Initial bits of documentation to get started and kick this off.

* Reverted back to main branch version

* Rewrote motivation to index

Per suggestions in PR #9, the motivation.md file content has been moved to the index.md file and rewritten to both be more document storage-format agnostic and place an empathis on digidal garden creation earlier on. 

I also expanded it with some general ideas on the vision of where NoteMate could be expanded to.

* Delete settings.json

Should not be part of any commits

* second part did not stage into commit

go figure. Workflow is a bit unstable today.