Docstrings for files (#11)
* Initial documentation bits

Initial bits of documentation to get started and kick this off.

* Cleaned up docstrings branch

* Added docstrings to rest of franklin.jil

Rolled the improved docstrings by @TheCedarPrince into the existing docstrings, and added docstrings of the same style to the rest of the methods. 

Docstring for generate_citation() still requires a clarification for the citations keyword argument - structure of the agument data unclear.

* Added docsstrings for remaining functions

We should now have docstring coverage across the entire NoteMate module.