Tagged Release for 0.0.1 Iteration (#30)
* Documentation coverage (#9)

* Initial documentation bits

Initial bits of documentation to get started and kick this off.

* Reverted back to main branch version

* Rewrote motivation to index

Per suggestions in PR #9, the motivation.md file content has been moved to the index.md file and rewritten to both be more document storage-format agnostic and place an empathis on digidal garden creation earlier on. 

I also expanded it with some general ideas on the vision of where NoteMate could be expanded to.

* Delete settings.json

Should not be part of any commits

* second part did not stage into commit

go figure. Workflow is a bit unstable today.

* Revision of docs landing page (#13)

* Revision of docs landing page

* Minor refactor for initial draft

* Docstrings for files (#11)

* Initial documentation bits

Initial bits of documentation to get started and kick this off.

* Cleaned up docstrings branch

* Added docstrings to rest of franklin.jil

Rolled the improved docstrings by @TheCedarPrince into the existing docstrings, and added docstrings of the same style to the rest of the methods. 

Docstring for generate_citation() still requires a clarification for the citations keyword argument - structure of the agument data unclear.

* Added docsstrings for remaining functions

We should now have docstring coverage across the entire NoteMate module.

* [DOCS] Add API Reference Page (#17)

* Added LiveServer dependency for easier local documentation previews

* Added link to API reference page

* Added API reference section

* Added first draft of README (#20)

* [DOCS] Example Workflow Showing What `NoteMate.jl` Can Do (#26)

* Writing initial usage documentation

Using the example workflow Julia parser, wrote a how to use document. Not yet hyperlinked to any other pages.

* Stubbing out parsing component

* Drafting first pass of NoteMate structure document

* Expanding using notemate tutorial

* Finished initial pass through and added todos

* Added tutorial section

* Finalized initial tutorial

* Added resources for documentation

* Updated package versions and added Markdown dependency

---------

Co-authored-by: Sevoris Doe <74614227+SevorisDoe@users.noreply.github.com>

* [FEATURE] Added Templates and Changelog (#29)

* Added templates for feature requests and bug reports

* Updated installation instructions

* Drafted initial Changelog

* Added a pull request template for future contributors

* Added tag for initial release

* Removed additional dependenices for initial package release

* Added test dependencies

* Added first attempt at a test suite for package

* Test assets for the test suite

---------

Co-authored-by: Sevoris Doe <74614227+SevorisDoe@users.noreply.github.com>